chore: improve token decimal conversion
