+++ common/token-module/src/lib.rs
@@ -2,7 +2,6 @@
 
 use klever_sc::derive_imports::*;
 use klever_sc::imports::*;
-use dfp_big_uint::DFPBigUint;
 
 pub const PERCENTAGE_TOTAL: u32 = 10_000; // precision of 2 decimals
 pub static INVALID_PERCENTAGE_SUM_OVER_ERR_MSG: &[u8] = b"Percentages do not add up to 100%";
@@ -130,26 +129,26 @@ pub trait TokenModule: fee_estimator_module::FeeEstimatorModule
         self.token_whitelist().swap_remove(&token_id);
     }
 
+    /// Returns true if tokens were successfully minted/released
+    /// Returns false if amount exceeds available balance or token is not whitelisted
+    /// Note: Expects amount already converted to KDA decimals
     #[endpoint(getTokens)]
-    fn get_tokens(&self, token_id: &TokenIdentifier, eth_amount: &BigUint) -> bool {
+    fn get_tokens(&self, token_id: &TokenIdentifier, kda_amount: &BigUint) -> bool {
         let caller = self.blockchain().get_caller();
         require!(
             caller == self.multi_transfer_contract_address().get(),
             "Only MultiTransfer can get tokens"
         );
 
-        // Convert from Ethereum decimals to Klever decimals
-        let kda_amount = self.convert_eth_to_kda_amount(token_id, eth_amount);
-
         if !self.mint_burn_token(token_id).get() {
             let total_balances_mapper = self.total_balances(token_id);
-            if &total_balances_mapper.get() >= &kda_amount {
+            if &total_balances_mapper.get() >= kda_amount {
                 total_balances_mapper.update(|total| {
-                    *total -= &kda_amount;
+                    *total -= kda_amount;
                 });
                 self.tx()
                     .to(ToCaller)
-                    .single_kda(token_id, 0, &kda_amount)
+                    .single_kda(token_id, 0, kda_amount)
                     .transfer();
 
                 return true;
@@ -162,49 +161,25 @@ pub trait TokenModule: fee_estimator_module::FeeEstimatorModule
         let mint_balances_mapper = self.mint_balances(token_id);
         if self.native_token(token_id).get() {
             require!(
-                burn_balances_mapper.get() >= &mint_balances_mapper.get() + &kda_amount,
+                burn_balances_mapper.get() >= &mint_balances_mapper.get() + kda_amount,
                 "Not enough burned tokens!"
             );
         }
 
-        let mint_executed = self.internal_mint(token_id, &kda_amount);
+        let mint_executed = self.internal_mint(token_id, kda_amount);
         if !mint_executed {
             return false;
         }
         self.tx()
             .to(ToCaller)
-            .single_kda(token_id, 0, &kda_amount)
+            .single_kda(token_id, 0, kda_amount)
             .transfer();
 
         mint_balances_mapper.update(|minted| {
-            *minted += &kda_amount;
+            *minted += kda_amount;
         });
 
         true
-    }
-
-    /// Convert amount from Ethereum decimals to Klever decimals
-    /// Uses stored decimals for both ETH and KDA sides
-    fn convert_eth_to_kda_amount(&self, token_id: &TokenIdentifier, eth_amount: &BigUint) -> BigUint {
-        let eth_decimals_mapper = self.eth_token_decimals(token_id);
-        let kda_decimals_mapper = self.kda_token_decimals(token_id);
-        
-        require!(
-            !eth_decimals_mapper.is_empty(),
-            "ETH decimals not configured for this token. Call setTokenDecimals first."
-        );
-        
-        require!(
-            !kda_decimals_mapper.is_empty(),
-            "KDA decimals not configured for this token. Call setTokenDecimals first."
-        );
-        
-        let eth_decimals = eth_decimals_mapper.get();
-        let kda_decimals = kda_decimals_mapper.get();
-        
-        DFPBigUint::from_raw(eth_amount.clone(), eth_decimals)
-            .convert(kda_decimals)
-            .to_raw()
     }
 
     #[only_admin]
