+++ kda-safe/wasm/src/lib.rs
@@ -5,8 +5,8 @@
 
 // Init:                                 1
 // Upgrade:                              1
-// Endpoints:                           57
-// Total number of exported functions:  59
+// Endpoints:                           58
+// Total number of exported functions:  60
 
 #![no_std]
 
@@ -31,6 +31,7 @@ klever_sc_wasm_adapter::endpoints! {
         getTotalRefundAmounts => get_total_refund_amounts
         getRefundFeesForEthereum => get_refund_fees_for_ethereum
         getTransactionFees => get_transaction_fees
+        convertEthToKdaAmount => convert_eth_to_kda_amount_endpoint
         getBridgedTokensWrapperAddress => bridged_tokens_wrapper_address
         setFeeEstimatorContractAddress => set_fee_estimator_contract_address
         setEthTxGasLimit => set_eth_tx_gas_limit
